add non interactive spinner
