+++ lib/CLI/CliUx.js
@@ -4,11 +4,12 @@
  * can be found in the LICENSE file.
  */
 var Table         = require('cli-table2');
-var Common        = require('../Common');
 var p             = require('path');
-var UX            = module.exports = {};
+var chalk         = require('chalk');
+var Common        = require('../Common');
+var Spinner       = require('./Spinner.js');
 
-var chalk = require('chalk');
+var UX = module.exports = {};
 
 /**
  * Description
@@ -269,54 +270,6 @@ UX.dispKeys = function(kv, target_module) {
     }
   });
 }
-
-var defaultSpinnerString = [
-  "⠋",
-  "⠙",
-  "⠹",
-  "⠸",
-  "⠼",
-  "⠴",
-  "⠦",
-  "⠧",
-  "⠇",
-  "⠏"
-].join('');
-
-var Spinner = function(textToShow){
-  this.text = textToShow || '';
-  this.setSpinnerString(defaultSpinnerString); // use default spinner string
-};
-
-Spinner.setDefaultSpinnerString = function(value) {
-  defaultSpinnerString = value;
-};
-
-Spinner.prototype.start = function() {
-  var current = 0;
-  var self = this;
-  this.id = setInterval(function() {
-    try {
-      process.stdout.clearLine();
-      process.stdout.cursorTo(0);
-      process.stdout.write(self.chars[current] + ' ' + self.text);
-      current = ++current % self.chars.length;
-    } catch(e) { // ignore error if term is not tty, just display nothing
-    }
-  }, 80);
-};
-
-Spinner.prototype.setSpinnerString = function(str) {
-  this.chars = str.split("");
-};
-
-Spinner.prototype.stop = function() {
-  try {
-    process.stdout.clearLine();
-    process.stdout.cursorTo(0);
-  } catch(e) {}
-  clearInterval(this.id);
-};
 
 UX.processing = {
   current_spinner : null,
